requirements.txt
websockets==12.0
python-dotenv==1.0.1
ccxt==4.4.31
scikit-learn>=1.3.0
xgboost>=2.0.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0
